Develop (#54)
* Fix/bug description (#26)

* 修复敏感词规则中无法添加多个敏感词的问题，新增规则增加编辑按钮

* 新增玩家举报板块，玩家可在发帖时设置是否匿名发帖，管理员后台增加自定义板块。

* Fix/bug description (#29)

* 修复敏感词规则中无法添加多个敏感词的问题，新增规则增加编辑按钮

* 新增玩家举报板块，玩家可在发帖时设置是否匿名发帖，管理员后台增加自定义板块。

* 更新用户可上传文件类型，可上传文件大小从20MB提升至50MB (#30)

* Fix/file upload error (#31)

* 更新用户可上传文件类型，可上传文件大小从20MB提升至50MB

* 修复表格列未对齐导致CI自动化流程出现的问题

* 规范化 claimed MIME + 兼容浏览器常见别名

* Fix/misc bugs (#34)

* 修复逗号被识别为违禁词，违禁词规则无法删除的问题

* 修复逗号被识别为违禁词，违禁词规则无法停用，违禁词规则无法被删除的问题

* 治理中心新增编辑删除等功能

* 修复用户无法使用steam登录的问题

* 修复CI自动化流程错误问题

* Fix/misc bugs (#36)

* 修复逗号被识别为违禁词，违禁词规则无法删除的问题

* 修复逗号被识别为违禁词，违禁词规则无法停用，违禁词规则无法被删除的问题

* 治理中心新增编辑删除等功能

* 修复用户无法使用steam登录的问题

* 修复CI自动化流程错误问题

* workers 代码版本检测

* workers更新，该项目与LumiAdmin项目共用worker

* feat: 评论区支持上传图片与附件

评论编辑器新增上传图片（comment_image，≤8MB）与附件（attachment，≤50MB）功能，
上传成功后自动在光标处插入 Markdown 链接；FileUpload 组件新增 compact 紧凑模式。

* style: 修复 comment-section.tsx 的 Prettier 格式

* security: 认证限流、反代 IP 支持、图片处理移出 async 运行时、角色判定统一

P0 安全修复：
- auth: 登录/注册/刷新添加基于 Redis 的速率限制（按 IP + 登录标识符），
  防暴力破解/撞库/注册轰炸；Redis 不可用时放行（不阻断用户）
- client_ip: 新增 TRUST_PROXY 中间件，从 X-Forwarded-For 解析真实客户端
  IP 并覆盖 ConnectInfo，修复反代后面限流/审计 IP 全为代理地址的问题
- upload: 图片解码/缩放/编码包 spawn_blocking，避免阻塞 tokio worker
- 统一前端角色判定为 @lumiforum/shared 的 ROLES/isElevatedRole/isAdminRole，
  修复 senior_moderator 在编辑器与详情页权限不一致的缺陷
- admin: 修复 update_role_permissions 返回 role_name 错误填充为 code 的 bug
- ci: rust job 接入 PostgreSQL/Redis services，集成测试不再静默跳过

* docs: 明确 TRUST_PROXY 与 X-Forwarded-For 信任要求

* feat: P1 优化——论坛页 RSC 化、CSRF 全局覆盖、401 联动、修改密码

- RSC 化：首页/板块页/话题详情页服务端取数，通过 initialData 注入客户端
  React Query 缓存，消除首屏双份 API 请求；非法 slug 返回真正 404
- CSRF：全局非 GET 路由加 enforce_mutation_origin；缺失 Origin 放行
  （浏览器跨站必带 Origin，curl/服务端客户端可继续使用 API）
- 401 联动：optionalAuthHeaders 过期请求触发刷新重试；刷新失败广播
  session-expired 事件，AuthProvider 暴露 sessionExpired 状态，
  论坛布局显示可关闭的登录过期横幅；API 请求添加 30s 超时
- 修改密码：POST /users/profile/password 端点 + 个人中心设置密码表单
  （Steam-only 账户可直接设置，已有密码账户须验证当前密码）
- 新增根级 not-found.tsx

* feat: P2 优化——metrics 认证、refresh token 清理、投票单事务、死缓存清理、CSP、移动导航、用户主页、前端测试

- /metrics 端点加 moderation.metrics.read 权限保护（不再公开暴露）
- 新增 refresh_tokens 每日过期清理任务（防止表无界增长）
- 投票 vote_many 单事务化：一次锁 poll 行 + 全部选项原子提交，
  消除多选投票 N 次独立事务与锁竞争
- 删除 Reaction 只写不读的死缓存（stats:topic/comment/user），
  移除每次点赞/收藏/关注的无效 Redis 往返
- Web 添加 CSP + Permissions-Policy 安全头（connect-src 允许 API/WS 域）
- 新增 Dependabot 配置（npm/cargo/github-actions 周度更新）
- 移动端导航菜单（汉堡按钮 + 下拉面板）
- 用户主页 GET /users/{id} 端点 + /users/[id] 页面（RSC 取数 + 关注交互），
  通知/搜索/JSON-LD 链接指向主页
- 前端测试：node:test 原生 TS 测试（SEO utils + zod schema 11 个用例），
  修复 plainText 链接文本替换 bug，test script 不再是 echo

* feat: 私信、暗色模式、@提及、草稿保存、话题标签

- 私信：conversations/messages 表与 dm 权限迁移；DM 服务/路由
  （按用户名寻址、软删、未读计数、收信通知）；/messages 双栏
  会话页、头部信封入口、用户主页发私信按钮
- 暗色模式：ThemeProvider + .dark 变量 + class 切换，补齐
  Input/Textarea/Select/头部等 bg-white 硬编码，根布局防闪烁脚本
- @提及：话题正文也触发提及通知（评论已有），修正通知文案
- 草稿保存：发帖草稿 localStorage 自动保存/恢复/清除
- 话题标签：编辑器标签输入 + 详情页徽章展示（后端表与归一化）

* style: 修复 Prettier 格式（私信/草稿/标签相关前端文件）

* fix+feat: 暗色对比度修复、作者资料链接、头像大图、私聊按钮改进

- 暗色模式：bg-foreground+text-white 组合（品牌标/置顶徽章）在
  深色下变为白底白字，改用 bg-primary+text-primary-foreground；
  Button 的 zinc 硬编码边框/悬停色改用语义 token
- 帖子列表与帖子页作者名、评论区用户名均改为个人资料页链接
- 新增 AvatarPreview：点击头像弹出大图灯箱（Esc/点击遮罩关闭），
  应用于帖子页与评论头像
- 个人资料页「发私信」更名为「私聊」，失败时显示错误信息，
  自己的资料页不显示（沿用 canFollow 判定）

* fix: 个人中心导航、用户主页帖子过滤、私信离线可用

- /profile 头部加「返回首页/我的收藏/通知中心」导航
- server.ts fetchTopics 之前漏传 author_id，导致 /users/[id] 服务端
  预取的是全站帖子（initialData 展示错误数据）；补上后客户端
  refetch 也会对齐
- 私信：unread 统计查询容错（缺失行不再 500）；会话头显示对方
  在线状态，离线时提示「消息会在对方上线后送达」，发送不受影响；
  修正 deleteMessage 的 URL 与后端路由一致
- 新增迁移补齐 notifications_type_check 缺失的 poll_voted/poll_ended
  （moderation 迁移重写约束时投票类型尚未加入，投票通知一直落库失败）

* fix: /messages 500 —— unread 计数 int4 解码为 i64 失败

conversations.user1_unread/user2_unread 是 INT4，而会话列表/未读统计
把它按 i64 解码（sqlx 严格类型匹配：INT4→i32、INT8→i64），导致每次
GET /messages 都在行解码阶段报错并返回 500。在三处 SQL 里显式
::bigint 后与 i64 匹配：list_for_user、unread_for、total_unread。

+++ apps/api/src/repositories/dm.rs
@@ -146,7 +146,7 @@ impl ConversationRepository {
                 ou.avatar_url AS other_avatar,
                 r.code AS other_role_code,
                 r.name AS other_role_name,
-                CASE WHEN c.user1_id = $1 THEN c.user1_unread ELSE c.user2_unread END AS unread_count,
+                CASE WHEN c.user1_id = $1 THEN c.user1_unread ELSE c.user2_unread END::bigint AS unread_count,
                 c.last_message_at,
                 (
                     SELECT m.content FROM messages m
@@ -341,7 +341,7 @@ impl ConversationRepository {
             r#"
             SELECT COALESCE(SUM(
                 CASE WHEN user1_id = $1 THEN user1_unread ELSE user2_unread END
-            ), 0)
+            ), 0)::bigint
             FROM conversations
             WHERE user1_id = $1 OR user2_id = $1
             "#,
@@ -359,7 +359,7 @@ impl ConversationRepository {
     ) -> Result<i64, sqlx::Error> {
         sqlx::query_scalar::<_, i64>(
             r#"
-            SELECT CASE WHEN user1_id = $2 THEN user1_unread ELSE user2_unread END
+            SELECT CASE WHEN user1_id = $2 THEN user1_unread ELSE user2_unread END::bigint
             FROM conversations
             WHERE id = $1 AND ($2 = user1_id OR $2 = user2_id)
             "#,
